Extract paged specific editor routes and views
