+++ entry_types/paged/packages/pageflow-paged/src/editor/index.js
@@ -1,4 +1,4 @@
-import {editor, app} from 'pageflow/editor';
+import {editor} from 'pageflow/editor';
 
 import {PagedEntry} from './models/PagedEntry';
 
@@ -7,6 +7,20 @@ import {EntryOutlineView} from './views/EntryOutlineView';
 import {EntryPreviewView} from './views/EntryPreviewView';
 import {appearanceInputs} from './helpers/appearanceInputs';
 import './views/PageSelectionView';
+
+import {SideBarRouter} from './routers/SideBarRouter';
+import {SideBarController} from './controllers/SideBarController';
+
+import './views/configurationEditors/audio';
+import './views/configurationEditors/backgroundImage';
+import './views/configurationEditors/video';
+import './views/configurationEditors/groups/background';
+import './views/configurationEditors/groups/general';
+import './views/configurationEditors/groups/pageLink';
+import './views/configurationEditors/groups/pageTransitions';
+import './views/configurationEditors/groups/options';
+
+export * from './views/inputs/PageLinkInputView';
 
 export * from 'pageflow/editor';
 export * from './models/PreviewEntryData';
@@ -17,4 +31,9 @@ editor.registerEntryType('paged', {
   previewView: EntryPreviewView,
   outlineView: EntryOutlineView,
   appearanceInputs: appearanceInputs
+});
+
+editor.registerSideBarRouting({
+  router: SideBarRouter,
+  controller: SideBarController
 });
