+++ entry_types/paged/packages/pageflow-paged/src/editor/templates/editChapter.jst
