+++ entry_types/paged/packages/pageflow-paged/src/editor/views/configurationEditors/backgroundImage.js
@@ -1,6 +1,6 @@
 import {ConfigurationEditorView} from 'pageflow/ui';
 
-import {FileInputView} from '../inputs/FileInputView';
+import {FileInputView} from 'pageflow/editor';
 
 import {state} from '$state';
