+++ entry_types/paged/packages/pageflow-paged/src/editor/views/configurationEditors/groups/pageTransitions.js
