+++ packages/pageflow/src/editor/controllers/SidebarController.js
@@ -5,12 +5,8 @@ import {EntryPublication} from '../models/EntryPublication';
 import {editor} from '../base';
 
 import {ConfirmEncodingView} from '../views/ConfirmEncodingView';
-import {EditChapterView} from '../views/EditChapterView';
 import {EditEntryView} from '../views/EditEntryView';
 import {EditMetaDataView} from '../views/EditMetaDataView';
-import {EditPageLinkView} from '../views/EditPageLinkView';
-import {EditPageView} from '../views/EditPageView';
-import {EditStorylineView} from '../views/EditStorylineView';
 import {EditWidgetView} from '../views/EditWidgetView';
 import {FilesView} from '../views/FilesView';
 import {PublishEntryView} from '../views/PublishEntryView';
@@ -67,41 +63,6 @@ export const SidebarController = Marionette.Controller.extend({
     }));
 
     editor.setDefaultHelpEntry('pageflow.help_entries.publish');
-  },
-
-  storyline: function(id) {
-    this.region.show(new EditStorylineView({
-      model: this.entry.storylines.get(id)
-    }));
-  },
-
-  chapter: function(id) {
-    this.region.show(new EditChapterView({
-      model: this.entry.chapters.get(id)
-    }));
-  },
-
-  page: function(id, tab) {
-    var page = this.entry.pages.get(id);
-
-    this.region.show(new EditPageView({
-      model: page,
-      api: editor,
-      tab: tab
-    }));
-
-    editor.setDefaultHelpEntry(page.pageType().help_entry_translation_key);
-  },
-
-  pageLink: function(linkId) {
-    var pageId = linkId.split(':')[0];
-    var page = state.pages.getByPermaId(pageId);
-
-    this.region.show(new EditPageLinkView({
-      model: page.pageLinks().get(linkId),
-      page: page,
-      api: editor
-    }));
   },
 
   widget: function(id) {
