+++ packages/pageflow/src/editor/index.js
@@ -111,7 +111,6 @@ export * from './views/BackgroundPositioningSlidersView';
 export * from './views/OtherEntriesCollectionView';
 export * from './views/EditFileView';
 export * from './views/PageLinkConfigurationEditorView';
-export * from './views/EditPageView';
 export * from './views/BackButtonDecoratorView';
 export * from './views/FileSettingsDialogView';
 export * from './views/OtherEntryItemView';
@@ -129,8 +128,6 @@ export * from './views/inputs/FileProcessingStateDisplayView';
 export * from './views/inputs/ReferenceInputView';
 export * from './views/inputs/ThemeInputView';
 export * from './views/inputs/FileInputView';
-export * from './views/inputs/PageLinkInputView';
-export * from './views/EditPageLinkView';
 export * from './views/EditEntryView';
 export * from './views/TextTracksView';
 export * from './views/FileThumbnailView';
@@ -144,8 +141,6 @@ export * from './views/DisabledAtmoIndicatorView';
 export * from './views/LockedView';
 export * from './views/ChangeThemeDialogView';
 export * from './views/ConfirmEncodingView';
-export * from './views/EditChapterView';
-export * from './views/EditStorylineView';
 export * from './views/FilteredFilesView';
 export * from './views/DropDownButtonItemView';
 export * from './views/EditMetaDataView';
@@ -164,15 +159,6 @@ export * from './views/FilesImporterView';
 export * from './views/ConfirmFileImportUploadView';
 
 export * from './views/EditConfigurationView';
-
-import './views/configurationEditors/audio';
-import './views/configurationEditors/backgroundImage';
-import './views/configurationEditors/video';
-import './views/configurationEditors/groups/background';
-import './views/configurationEditors/groups/general';
-import './views/configurationEditors/groups/pageLink';
-import './views/configurationEditors/groups/pageTransitions';
-import './views/configurationEditors/groups/options';
 
 import './views/widgetTypes/classicLoadingSpinner';
 import './views/widgetTypes/cookieNoticeBar';
